fix(frontend): Reduce width of the Color Map
The Color Map is similar in nature to the vector Color Range slider.
The Color Map end value was being hidden due to the widget being too
long. I have made it so the width of the widget, is less than 100%,
which seems to have fixed the problem, even with scrollbars in place.