fix(frontend): Assign default MapStyles to new Customers
When a new customer is created, we need to assign at least one MapStyle
to it, so that the app doesn't break, once the user has logged in.
Now, each MapStyles has a flag `is_default` and the MapStyles marked
as True, are assigned to the customer. If the customer needs non default
MapStyles, that is an admin task and must be done manually.

IssueID ORB-1370