fix(frontend): Improve quality of test
Added missing mandatory fields to the factory for MapStyle objects.

Improved test itself so that we test styles that shouldn't exist,
don't and that those that should, do.

I was using a LicencedCustomer, but after comments in review, I changed
it as it was unnecessary, a straight-forward Customer, is enough.

IssueID ORB-1370