fix(frontend): Separate data layer/crossfiltering state
The global state used by Data Loading and CrossFiltering should be
separated, so we can reset back to each, if we switch between these
two functions. This is a fair bit of duplication, just keeping the
state as siblings of each other, but it keeps a good separation of
concern.

IssueID ORB-1415