feat(backend): Add abilty to backup/restore server dbase
On occasion (and for new devs) we need to re-create the base, usually
when we have lost the dock postgres container, through human error
etc. I added the `django-dbbackup` django app, to add a mangagement
command that we can export the postgres db as SQL and then restore it
to another instance.