fix(frontend): Make crossfilterable sources global state
Instead of calculating the list of cross-filterable data sources
on-the-fly, I now calculate this and store in global state at the
same time we successfully fetch the whole list. For a little extra
state, we save effort later on, retrieving this list as it is likely
to be needed in multiple places, so figured this would be easiest.

IssueID ORB-1415