feat(frontend): Add cross-filtering layer redux handling
This was a copy of the `layers.slice.js` and then reduced down to
just the code necessary for cross-filtering. It is possible that we
will move all redux code for cross-filtering here eventually, but
for now, I've got some elsewhere as it copies what we have for normal
data loading.

IssueID ORB-1418