feat(frontend): Add new cross-filtering layer config
Using the ISO+ data layer config, I have simplified it a lot to just
get the cross-filtering data visible on the map. I fully expect more
work to come on this, to add `Data Filtering` for instance, as well
as set the colours of map features correctly, but that can easily be
copied from `isolationPlusLayerConfig.js`, when we focus on that.

IssueID ORB-1418