feat(frontend): Get right layers, sidebar and map components
The `useOrb` hook is very specific to normal Data Loading. I will be
extending it to work for cross-filtering, as well as normal Data Loading.
That means passing some config to the hook, so this code it to get the
right data, based on whether we are cross-filtering or not and passing
it to the hook.

IssueID ORB-1418