fix(frontend): Add state management for cross-filtering common geometry
When in cross-filtering mode, we need to figure out the common geometry
to be used. Right now, since I'm waiting for other code to do this to be
written, I'm hard-coding it, along with the sourceId's of the selected
cross-filterable layers. The hard-coded state will be replaced when the
other code is written.

IssueID ORB-1418