feat(frontend): Disable checkboxes once four selected
add check to monitor length of selectedProperties. Checkboxes
are disabled once 4 are selected, but selected ones are still enabled
so user is able to choose a different 4th option by deselection
one or more of them. No lower check for 2, that will be done in
next commit and on the Confirm button

IssueID ORB-1421