feat(frontend): create unit tests for getGeometryType and groupPropertiesAndSourceIds finctions in the filter-layer-view.componet.js
I extracted two functions from the filter-layer-view.componet.js, so I can export them. There are two tests for the getGeometryType and three tests for the groupPropertiesAndSourceIds function. In the last test (check that function returns an empty object when dataSources undefined), I expect the function to pass the test when const expected is equal to an empty object; in other cases, the test should fail. When I set const expected = { test1: undefined }, the test passed, but I think that it should fail. I don't understand why this happens.

IssueId: ORB-1423