feat(frontend): Add Data filtering of the map data
Added the ability to take the slider values in the sidebar and use
them to filter the data on the map.

I've hard-coded the list of cross-filter property/values, as these
all need set at the start.

It isn't quite as fast I have seen, going to look at that next.