feat(frontend): change the order of geometry type and variables names
I changed the order of geometry type from lowest to highest, and I updated my test according to that changes.
I the data-layers-slice.js file, I set crossFilteringCommonGeometry to an empty string. When I try to set a value equal to null, and when I click on the cross-filtering icon, there is an error in my browser referring to that null value. I renamed some variables and files.

IssueID ORB-1423