fix(frontend): Refactor dialog submit common geometry, remove log
Refactored the dialog submit inside filter-layer-view to use the commonGeometry calculated inside
the submit handler, as it has not yet been dispatched to global state, therefore the selector value
will always be null and cannot be used.

The value that is eventually dispatched is calculeted inside that same dubmit handler, so it's easy
to just grab it from there.

Also removed a stray console log.