fix(frontend): Flip selecteed geometry from [length - 1] to [0]
This flips the selected geometry type to be the most granular, so closer to OA rather than LAD_2020.

Once all of the selected properties have identified their lowest geometry available, they are all
fed into the function that determines the highest of those (closest to LAD_2020) and selects that
as the level for all to be aggregated up to.