fix(frontend): Improve tooltips
I've improved the content of the tooltip info buttons in cross-filtering
for the Dialog, where we display the property description along with the
dataset description and the granularity of the data. In the sidebar, for
cross-filtering, we combine just the property and dataset descriptions,
as the data is aggregated and to show the info from metadata would be
incorrect.

I also added the category path to the sidebar info, so you know which
categories the data is from.