feat(frontend): Rename variables, selectors, actions, types
Renamed all of the variables, selectors, actions and types relating to the CF slice and its tests,
the layer config, the dialog picker, and the sidebar radio picker.

The action to set the initial values now has `initialValues` in the name, and the actions/selectors/variables
relating to `filterValue` or `clipValue` are now called `filterRange` and `clipRange`.

This gives a much more explicit description of what is actually being passed around, making it less
confusing for ourselves and others to follow through the app.y